Enhancement and Bug Fixes
Added ImageUploadService, Fixed DeleteImage Bug, Added DeleteTeam Confirmation

+++ src/service/teamsHandler/deleteTeamHandler.jsx
@@ -5,18 +5,22 @@ import deleteImageHandler from "../deleteImageService";
 const deleteTeamHandler = async (team) => {
   const { teamLogoURL, teamBannerURL, teamName } = team;
 
-  try {
-    axios.delete(`${appSettings.teams}/${teamName}`).then(() => {
-      alert("Deleted from Firebase Firestore/Storage successfully!");
-    });
+  const confirmed = window.confirm(
+    `Are you sure you want to delete the team ${teamName}?`
+  );
+
+  if (confirmed) {
     await Promise.all([
       deleteImageHandler(teamLogoURL),
       deleteImageHandler(teamBannerURL),
     ]);
 
+    await axios.delete(`${appSettings.teams}/${teamName}`).then(() => {
+      console.log("Deleted from Firebase Firestore successfully!");
+    });
+
+    alert("Deleted from Firestore/Storage successfully!");
     window.location.reload();
-  } catch (err) {
-    alert(`Unable to delete Team`);
   }
 };
